refactor: execMintStatemine and execMintStatemine
